+++ lib/chef/knife/winrm.rb
@@ -31,13 +31,6 @@ class Chef
       include Chef::Knife::WinrmCommandSharedFunctions     
       include Chef::Knife::KnifeWindowsBase
 
-      FAILED_BASIC_HINT ||= "Hint: Please check winrm configuration 'winrm get winrm/config/service' AllowUnencrypted flag on remote server."
-      FAILED_NOT_BASIC_HINT ||= <<-eos.gsub /^\s+/, ""
-        Hint: Make sure to prefix domain usernames with the correct domain name.
-        Hint: Local user names should be prefixed with computer name or IP address.
-        EXAMPLE: my_domain\\user_namer
-      eos
-
       deps do
         require 'readline'
         require 'chef/search/query'
@@ -60,64 +53,12 @@ class Chef
       end
 
       def execute_remote_command
-        begin
-          case @name_args[1]
-          when "interactive"
-            interactive
-          else
-            relay_winrm_command(@name_args[1..-1].join(" "))
-
-            if config[:returns]
-              check_for_errors!
-            end
-
-            # Knife seems to ignore the return value of this method,
-            # so we exit to force the process exit code for this
-            # subcommand if returns is set
-            exit @exit_code if @exit_code && @exit_code != 0
-            @exit_code || 0
-          end
-        rescue WinRM::WinRMHTTPTransportError, WinRM::WinRMAuthorizationError => e
-          if authorization_error?(e)
-            if ! config[:suppress_auth_failure]
-              # Display errors if the caller hasn't opted to retry
-              ui.error "Failed to authenticate to #{@name_args[0].split(" ")} as #{locate_config_value(:winrm_user)}"
-              ui.info "Response: #{e.message}"
-              ui.info get_failed_authentication_hint
-              raise e
-            end
-            @exit_code = 401
-          else
-            raise e
-          end
+        case @name_args[1]
+        when "interactive"
+          interactive
+        else
+          run_command(@name_args[1..-1].join(" "))
         end
-      end
-
-      def relay_winrm_command(command)
-        Chef::Log.debug(command)
-        @winrm_sessions.each do |s|
-          s.relay_command(command)
-        end
-      end
-
-      # TODO: Copied from Knife::Core:GenericPresenter. Should be extracted
-      def extract_nested_value(data, nested_value_spec)
-        nested_value_spec.split(".").each do |attr|
-          if data.nil?
-            nil # don't get no method error on nil
-          elsif data.respond_to?(attr.to_sym)
-            data = data.send(attr.to_sym)
-          elsif data.respond_to?(:[])
-            data = data[attr]
-          else
-            data = begin
-                     data.send(attr.to_sym)
-                   rescue NoMethodError
-                     nil
-                   end
-          end
-        end
-        ( !data.kind_of?(Array) && data.respond_to?(:to_hash) ) ? data.to_hash : data
       end
 
       private
@@ -152,16 +93,6 @@ class Chef
         end
       end
 
-      def check_for_errors!
-        @winrm_sessions.each do |session|
-          session_exit_code = session.exit_code
-          unless success_return_codes.include? session_exit_code.to_i
-            @exit_code = session_exit_code.to_i
-            ui.error "Failed to execute command on #{session.host} return code #{session_exit_code}"
-          end
-        end
-      end
-
       # Present the prompt and read a single line from the console. It also
       # detects ^D and returns "exit" in that case. Adds the input to the
       # history, unless the input is empty. Loops repeatedly until a non-empty
@@ -186,26 +117,6 @@ class Chef
       def reader        
         Readline
       end
-
-      def authorization_error?(exception)
-        exception.is_a?(WinRM::WinRMAuthorizationError) ||
-          exception.message =~ /401/
-      end
-
-      def success_return_codes
-        #Redundant if the CLI options parsing occurs
-        return [0] unless config[:returns]
-        return @success_return_codes ||= config[:returns].split(',').collect {|item| item.to_i}
-      end
-
-      def get_failed_authentication_hint
-        if @session_opts[:basic_auth_only]
-          FAILED_BASIC_HINT
-        else
-          FAILED_NOT_BASIC_HINT
-        end
-      end
     end
   end
 end
-