+++ spec/unit/knife/bootstrap_windows_winrm_spec.rb
@@ -28,19 +28,26 @@ describe Chef::Knife::BootstrapWindowsWinrm do
   before do
     bootstrap.config[:run_list] = []
     allow(bootstrap).to receive(:validate_options!).and_return(nil)
-    #    Kernel.stub(:sleep).and_return 10
     allow(bootstrap).to receive(:sleep).and_return(10)
+    allow(Chef::Knife::WinrmSession).to receive(:new).and_return(session)
     allow(File).to receive(:exist?).with(File.expand_path(Chef::Config[:validation_key])).and_return(true)
   end
 
   after do
-    #    Kernel.unstub(:sleep)
     allow(bootstrap).to receive(:sleep).and_return(10)
   end
 
-  let(:bootstrap) { Chef::Knife::BootstrapWindowsWinrm.new(['winrm', '-d', 'windows-chef-client-msi',  '-x', 'Administrator', 'localhost']) }
-  let(:session) { Chef::Knife::Winrm::WinrmSession.new({ :host => 'winrm.cloudapp.net', :port => '5986', :transport => :ssl }) }
-
+  let(:session_opts) do
+    {
+      user: "Administrator",
+      password: "testpassword", 
+      port: "5986",
+      transport: :ssl,
+      host: "localhost"
+    }
+  end
+  let(:bootstrap) { Chef::Knife::BootstrapWindowsWinrm.new(['winrm', '-d', 'windows-chef-client-msi',  '-x', session_opts[:user], '-P', session_opts[:password], session_opts[:host]]) }
+  let(:session) { Chef::Knife::WinrmSession.new(session_opts) } 
   let(:initial_fail_count) { 4 }
 
   context "knife secret-file && knife secret options are passed" do
@@ -135,6 +142,15 @@ describe Chef::Knife::BootstrapWindowsWinrm do
       expect(bootstrap.load_correct_secret).to eq(
         "data_bag_secret_key_passed_under_cli_secret_file_option")
     end
+  end
+
+  it 'should pass exit code from failed winrm call' do
+    allow(session).to receive(:exit_code).and_return(500)
+    allow(bootstrap).to receive(:wait_for_remote_response)
+    allow(bootstrap).to receive(:create_bootstrap_bat_command)
+    allow(session).to receive(:relay_command)
+    allow(bootstrap.ui).to receive(:info)
+    expect { bootstrap.run_with_pretty_exceptions }.to raise_error(SystemExit) { |e| expect(e.status).to eq(500) }
   end
 
   it 'should retry if a 401 is received from WinRM' do
@@ -237,6 +253,42 @@ describe Chef::Knife::BootstrapWindowsWinrm do
         expect(bootstrap.client_builder).not_to receive(:client_path)
         expect(bootstrap.ui).to receive(:error)
         expect { bootstrap.bootstrap }.to raise_error(SystemExit)
+      end
+    end
+  end
+
+  context "when doing chef vault", :chef_gte_12_only do
+    let(:vault_handler) { double('vault_handler', :doing_chef_vault? => true) }
+    let(:node_name) { 'foo.example.com' }
+    before do
+      allow(bootstrap).to receive(:wait_for_remote_response)
+      allow(bootstrap).to receive(:create_bootstrap_bat_command)
+      allow(bootstrap).to receive(:run_command).and_return(0)
+      bootstrap.config[:chef_node_name] = node_name
+      bootstrap.chef_vault_handler = vault_handler
+    end
+
+    context "builder does not respond to client" do
+      before do
+        bootstrap.client_builder = instance_double("Chef::Knife::Bootstrap::ClientBuilder", :run => nil, :client_path => nil)
+      end
+
+      it "passes a node search query to the handler" do
+        expect(vault_handler).to receive(:run).with(node_name: node_name)
+        bootstrap.bootstrap
+      end
+    end
+
+    context "builder responds to client" do
+      let(:client) { Chef::ApiClient.new }
+
+      before do
+        bootstrap.client_builder = double("Chef::Knife::Bootstrap::ClientBuilder", :run => nil, :client_path => nil, :client => client)
+      end
+
+      it "passes a node search query to the handler" do
+        expect(vault_handler).to receive(:run).with(client)
+        bootstrap.bootstrap
       end
     end
   end